# PERF: Improve read time for smaller num of assets.