@ssanderson 2016-04-21
LGTM pending your decision on the slicing change.